+++ ProyectoFinalAppi/Models/Producto.cs
@@ -3,12 +3,11 @@
     public class Producto
     {
         //Variables.
-        public long Id { get; set; }
-        public string Descripciones { get; set; } 
-        public decimal Costo { get; set; }
-        public decimal PrecioVenta { get; set; }
+        public int Id { get; set; }
+        public string Descripciones { get; set; }
+        public double Costo { get; set; }
+        public double PrecioVenta { get; set; }
         public int Stock { get; set; }
-        public long IdUsuario { get; set; }
-
+        public int IdUsuario { get; set; }
     }
 }
